refactor: seperate offset to bytes trait from transpose class
