+++ crates/proof-of-sql/src/proof_primitive/dory/mod.rs
@@ -138,6 +138,7 @@ type DeferredGT = deferred_msm::DeferredMSM<GT, F>;
 type DeferredG1 = deferred_msm::DeferredMSM<G1Affine, F>;
 type DeferredG2 = deferred_msm::DeferredMSM<G2Affine, F>;
 
+mod offset_to_bytes;
 mod pairings;
 mod transpose;
